rename controller concerns and add deprecated method to logger class
