+++ app/controllers/concerns/shopify_app/authenticated.rb
@@ -5,15 +5,10 @@ module ShopifyApp
     extend ActiveSupport::Concern
 
     included do
-      include ShopifyApp::Localization
-      include ShopifyApp::LoginProtection
-      include ShopifyApp::CsrfProtection
-      include ShopifyApp::EmbeddedApp
-      include ShopifyApp::EnsureBilling
-
-      before_action :login_again_if_different_user_or_shop
-      around_action :activate_shopify_session
-      after_action :add_top_level_redirection_headers
+      ShopifyApp::Logger.deprecated("RequireKnownShop has been renamed to EnsureInstalled."\
+        " Please use EnsureInstalled controller concern for the same behavior")
     end
+
+    include ShopifyApp::EnsureHasSession
   end
 end
