+++ app/controllers/concerns/shopify_app/require_known_shop.rb
@@ -3,54 +3,12 @@
 module ShopifyApp
   module RequireKnownShop
     extend ActiveSupport::Concern
-    include ShopifyApp::RedirectForEmbedded
 
     included do
-      if ancestors.include?(ShopifyApp::LoginProtection)
-        ActiveSupport::Deprecation.warn(<<~EOS)
-          We detected the use of incompatible concerns (RequireKnownShop and LoginProtection) in #{name},
-          which may lead to unpredictable behavior. In a future release of this library this will raise an error.
-        EOS
-
-      end
-
-      before_action :check_shop_domain
-      before_action :check_shop_known
+      ShopifyApp::Logger.deprecated("RequireKnownShop has been renamed to EnsureInstalled."\
+        " Please use EnsureInstalled controller concern for the same behavior")
     end
 
-    def current_shopify_domain
-      return if params[:shop].blank?
-
-      @shopify_domain ||= ShopifyApp::Utils.sanitize_shop_domain(params[:shop])
-    end
-
-    private
-
-    def check_shop_domain
-      redirect_to(ShopifyApp.configuration.login_url) unless current_shopify_domain
-    end
-
-    def check_shop_known
-      @shop = SessionRepository.retrieve_shop_session_by_shopify_domain(current_shopify_domain)
-      unless @shop
-        if embedded_param?
-          redirect_for_embedded
-        else
-          redirect_to(shop_login)
-        end
-      end
-    end
-
-    def shop_login
-      url = URI(ShopifyApp.configuration.login_url)
-
-      url.query = URI.encode_www_form(
-        shop: params[:shop],
-        host: params[:host],
-        return_to: request.fullpath,
-      )
-
-      url.to_s
-    end
+    include ShopifyApp::EnsureInstalled
   end
 end
