+++ lib/shopify_app/logger.rb
@@ -1,17 +1,16 @@
 # frozen_string_literal: true
 
 module ShopifyApp
-  module Logger
+  class Logger
     LOG_LEVELS = { debug: 0, info: 1, warn: 2, error: 3, off: 6 }
     DEFAULT_LOG_LEVEL = :info
 
     def self.send_to_logger(log_level, message)
       return unless enabled_for_log_level?(log_level)
 
-      current_shop = ShopifyAPI::Context.active_session&.shop || "Unknown Shop"
-      message_context = "[ ShopifyApp | #{log_level.to_s.upcase} | #{current_shop} ] #{message}"
+      full_message = "#{context(log_level)} #{message}"
 
-      ShopifyAPI::Context.logger.public_send(log_level, message_context)
+      ShopifyAPI::Context.logger.send(log_level, full_message)
     end
 
     def self.debug(message)
@@ -30,7 +29,20 @@ module ShopifyApp
       send_to_logger(:error, message)
     end
 
-    def self.enabled_for_log_level?(log_level)
+    def self.deprecated(message)
+      return unless enabled_for_log_level?(:warn)
+
+      ActiveSupport::Deprecation.warn("#{context(:warn)} #{message}")
+    end
+
+    private
+
+    def context(log_level)
+      current_shop = ShopifyAPI::Context.active_session&.shop || "Shop Not Found"
+      "[ ShopifyApp | #{log_level.to_s.upcase} | #{current_shop} ]"
+    end
+
+    def enabled_for_log_level?(log_level)
       raise(ShopifyApp::ConfigurationError,
         "Invalid Log Level - #{log_level}") unless LOG_LEVELS.keys.include?(log_level)
